Organized Files, Updated UI, and added DefaultDate
