+++ bin/.gitignore
@@ -1,1 +1,4 @@
 /application/
+/css/
+/images/
+/view/
