+++ resources/css/application.css
@@ -1,7 +1,6 @@
 .rounded-button {
     -fx-background-color: #1464f6;
     -fx-text-fill: white;
-    -fx-font-size: 25px;
     -fx-padding: 10px 20px;
     -fx-background-radius: 30;
     -fx-border-radius: 30;
@@ -14,14 +13,12 @@
 
 .rounded-cancel-button {
     -fx-background-color: #F55356;
-    -fx-text-fill: white;
-    -fx-font-size: 20px;
-    -fx-padding: 10px 20px;
-    -fx-background-radius: 30;
-    -fx-border-radius: 30;
-    -fx-border-color: transparent;
 }
 
 .rounded-cancel-button:hover {
     -fx-background-color: #c14446
+}
+
+.formButton{
+	-fx-font-size: 20px;
 }
