+++ resources/view/newAccount.fxml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import java.lang.String?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.DatePicker?>
 <?import javafx.scene.control.TextField?>
@@ -12,7 +13,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1080.0" xmlns="http://javafx.com/javafx/23" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.Controller">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1080.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.controller.AccountController">
 	<children>
 		<Rectangle arcHeight="5.0" arcWidth="5.0">
 			<fill>
@@ -43,9 +44,9 @@
 				<Font size="16.0" />
 			</font>
 		</TextField>
-		<Button layoutX="140.0" layoutY="560.0" mnemonicParsing="false" onAction="#switchToHome" styleClass="rounded-button" text="Home">
+		<Button layoutX="130.0" layoutY="537.0" mnemonicParsing="false" onAction="#switchToHome" styleClass="rounded-button" stylesheets="@../css/application.css" text="Home">
 			<font>
-				<Font size="50.0" />
+				<Font size="30.0" />
 			</font>
 		</Button>
 		<Text layoutX="444.0" layoutY="113.0" strokeType="OUTSIDE" strokeWidth="0.0" text="New Account" wrappingWidth="402.0">
@@ -53,15 +54,24 @@
 				<Font name="Lato Bold" size="50.0" />
 			</font>
 		</Text>
-		<Button fx:id="submit" layoutX="439.0" layoutY="490.0" onAction="#submitButton" prefHeight="60.0" prefWidth="224.0" styleClass="rounded-button" text="Submit" textFill="WHITE">
+		<Button fx:id="submit" layoutX="565.0" layoutY="485.0" onAction="#submitButton" prefHeight="40.0" prefWidth="110.0" stylesheets="@../css/application.css" text="Submit" textFill="WHITE">
 			<font>
-				<Font name="Lato Regular" size="28.0" />
+				<Font size="25.0" />
 			</font>
+         <styleClass>
+            <String fx:value="formButton" />
+            <String fx:value="rounded-button" />
+         </styleClass>
 		</Button>
-		<Button layoutX="438.0" layoutY="577.0" onAction="#switchToHome" prefHeight="43.0" prefWidth="116.0" styleClass="rounded-cancel-button" text="Cancel">
+		<Button fx:id="cancel" layoutX="444.0" layoutY="485.0" onAction="#switchToHome" prefHeight="29.0" prefWidth="110.0" stylesheets="@../css/application.css" text="Cancel">
 			<font>
-				<Font size="20.0" />
+				<Font size="50.0" />
 			</font>
+         <styleClass>
+            <String fx:value="rounded-cancel-button" />
+            <String fx:value="formButton" />
+            <String fx:value="rounded-button" />
+         </styleClass>
 		</Button>
 		<Text layoutX="446.0" layoutY="391.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Opening Balance (required)" wrappingWidth="261.0001220703125">
 			<font>
@@ -80,7 +90,7 @@
 		</Text>
 		<ImageView fitHeight="192.0" fitWidth="288.0" layoutX="47.0" layoutY="37.0" pickOnBounds="true" preserveRatio="true">
 			<image>
-				<Image url="@Budget%20Buddy%20Logo.png" />
+				<Image url="@../images/BudgetBuddyLogo.png" />
 			</image>
 		</ImageView>
 	</children>
